Killing the WorkerSupervisor restarts processes
In the last commit, killing the `WorkerSupervisor` would kill that
supervisor and all its workers, but the `Server` was still alive and
blissfully unaware. This problem is solved through linking the
`WorkerSupervisor` to the `Server` process on initialization. Now the
(correct, I believe) behavior is that killing either the `Server` or the
`WorkerSupervisor` will kill the other process as well. Then the
`Server` will be restarted by the `Supervisor` and it will create a new
`WorkerSupervisor` and its `Worker`s.

I wonder if maybe a better solution would be in the strategy choices at
the `Supervisor` level, but I'm not sure it would because the desired
restart behavior is a new `Server` etc.